updates entity/fetch-one to return nil when no document is found
Previously entity/fetch-one would return an empty entity (with default
values) when no matching document was found.  This is surprising as
it leads the caller to believe that a document was actually found and
coerced into the entity type.  By returning nil the caller can act
appropriately when no matching entity is found.